feat: view divided in steps
