+++ sprout/templates/projects-id-metadata-id-update.html
@@ -39,7 +39,7 @@
     </tbody>
   </table>
   <div class="row" style="justify-content: center">
-    <button type="submit" tabindex="10">Update All</button>
+    <button type="submit" tabindex="10">Continue</button>
   </div>
 </form>
 
@@ -102,7 +102,7 @@
   </div>
 
   <div class="row" style="justify-content: center">
-    <button type="submit" tabindex="10">Update All</button>
+    <button type="submit" tabindex="10">Continue</button>
   </div>
 </form>
